Add Base Class for CILViewer
